fix: schema argument type, linter issues
