+++ src/packages/form/src/validator/ExtentedSchema.ts
@@ -1,0 +1,12 @@
+/* eslint-disable import/no-extraneous-dependencies */
+import { JSONSchema7, JSONSchema7Definition } from 'json-schema';
+
+export interface ExtentedSchema extends JSONSchema7 {
+  properties?: {
+    [key: string]: JSONSchema7Definition & {
+      messages?: {
+        [key: string]: string | undefined;
+      };
+    };
+  };
+}
