+++ src/packages/ui/src/util/isElement.ts
@@ -1,6 +1,5 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-export function isElement(element: any): boolean {
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function isElement(element: HTMLElement | Element | any): boolean {
   if (typeof window === 'undefined') return false;
   return element instanceof Element || element instanceof HTMLElement;
 }
